Factor out project.util.success_url.
