+++ twofactor/tests/test_views.py
@@ -3,26 +3,11 @@ import pytest
 
 from users.tests.factories import UserFactory
 from .factories import TwofactorInfoFactory
-from twofactor.views import get_success_url
 
 
 SECRET_SENTINEL = 'otpauth://totp/'
 SCRIPT_SENTINEL = '</script>'
 ERROR_SENTINEL = 'Alas'
-
-
-class TestGetSuccessUrl:
-    def test_it_accepts_valid_get_params(self, rf):
-        req = rf.get('/', {'next': '/blah'})
-        assert get_success_url(req) == '/blah'
-
-    def test_it_accepts_valid_post_params(self, rf):
-        req = rf.post('/', {'next': '/blah'})
-        assert get_success_url(req) == '/blah'
-
-    def test_it_rejects_invalid_urls(self, rf):
-        req = rf.get('/', {'next': 'http://evil/thing'})
-        assert get_success_url(req) == '/'
 
 
 def test_verify_redirects_anonymous_users_to_login(client):
